requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.9.0